+++ dap-etl/transformation/src/test/scala/org/broadinstitute/monster/dap/CancerTransformationsSpec.scala
@@ -1,6 +1,6 @@
 package org.broadinstitute.monster.dap
 
-import org.broadinstitute.monster.dap.common.RawRecord
+import org.broadinstitute.monster.dap.common.{CancerTransformations, RawRecord}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
